backend/requirements.txt
# Snowflake and Data Processing Dependencies
snowflake-snowpark-python[pandas]==1.14.0
polars>=0.20.0
pandas>=2.0.0
pyarrow>=12.0.0
fastparquet>=0.8.0
psutil>=5.9.0

# Core Snowflake Connector (already used in existing code)
snowflake-connector-python>=3.0.0

# Redis for caching (already used in existing code)
redis>=4.5.0

# Environment variables management
python-dotenv>=1.0.0

# Anthropic API client
anthropic>=0.25.0

# Web framework for API endpoints
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Additional utilities
pydantic>=2.0.0
httpx>=0.24.0